requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
exceptiongroup==1.3.0
email-validator
Faker==37.12.0
Faker
greenlet==3.2.4
idna==3.11
iniconfig==2.1.0
packaging==25.0
playwright==1.56.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
requests==2.32.5
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
fastapi
uvicorn[standard]
jinja2
python-multipart



